+++ magpie/api/schemas.py
@@ -122,9 +122,6 @@ UserResourcePermissionsAPI = Service(
 UserResourcePermissionAPI = Service(
     path="/users/{user_name}/resources/{resource_id}/permissions/{permission_name}",
     name="UserResourcePermission")
-UserResourceTypesAPI = Service(
-    path="/users/{user_name}/resources/types/{resource_type}",
-    name="UserResourceTypes")
 UserServicesAPI = Service(
     path="/users/{user_name}/services",
     name="UserServices")
@@ -158,9 +155,6 @@ LoggedUserResourcePermissionAPI = Service(
 LoggedUserResourcePermissionsAPI = Service(
     path=LoggedUserBase + "/resources/{resource_id}/permissions",
     name="LoggedUserResourcePermissions")
-LoggedUserResourceTypesAPI = Service(
-    path=LoggedUserBase + "/resources/types/{resource_type}",
-    name="LoggedUserResourceTypes")
 LoggedUserServicesAPI = Service(
     path=LoggedUserBase + "/services",
     name="LoggedUserServices")
@@ -203,9 +197,6 @@ GroupResourcePermissionsAPI = Service(
 GroupResourcePermissionAPI = Service(
     path="/groups/{group_name}/resources/{resource_id}/permissions/{permission_name}",
     name="GroupResourcePermission")
-GroupResourceTypesAPI = Service(
-    path="/groups/{group_name}/resources/types/{resource_type}",
-    name="GroupResourceTypes")
 RegisterGroupsAPI = Service(
     path="/register/groups",
     name="RegisterGroups")
@@ -227,6 +218,9 @@ ResourceAPI = Service(
 ResourcePermissionsAPI = Service(
     path="/resources/{resource_id}/permissions",
     name="ResourcePermissions")
+ResourceTypesAPI = Service(
+    path="/resources/{resource_id}/types",
+    name="ResourceTypes")
 ServicesAPI = Service(
     path="/services",
     name="Services")
@@ -713,6 +707,14 @@ class ResourceTypesListSchema(colander.SequenceSchema):
     )
 
 
+class ChildrenResourceTypesListSchema(colander.SequenceSchema):
+    resource_type = colander.SchemaNode(
+        colander.String(),
+        description="Available children resource types under a resource within a service hierarchy.",
+        example="file",
+    )
+
+
 class GroupNamesListSchema(colander.SequenceSchema):
     group_name = GroupNameParameter
 
@@ -1147,7 +1149,7 @@ class Resources_GET_OkResponseSchema(BaseResponseSchemaAPI):
 class Resources_POST_RequestBodySchema(colander.MappingSchema):
     resource_name = colander.SchemaNode(
         colander.String(),
-        description="Name of the resource to create"
+        description="Name of the resource to create."
     )
     resource_display_name = colander.SchemaNode(
         colander.String(),
@@ -1156,11 +1158,11 @@ class Resources_POST_RequestBodySchema(colander.MappingSchema):
     )
     resource_type = colander.SchemaNode(
         colander.String(),
-        description="Type of the resource to create"
+        description="Type of the resource to create."
     )
     parent_id = colander.SchemaNode(
         colander.Int(),
-        description="ID of parent resource under which the new resource should be created",
+        description="ID of parent resource under which the new resource should be created.",
         missing=colander.drop
     )
 
@@ -1220,6 +1222,51 @@ class ResourcePermissions_GET_OkResponseSchema(BaseResponseSchemaAPI):
 class ResourcePermissions_GET_BadRequestResponseSchema(BaseResponseSchemaAPI):
     description = "Invalid resource type to extract permissions."
     body = ErrorResponseBodySchema(code=HTTPBadRequest.code, description=description)
+
+
+class ResourceTypes_GET_RequestSchema(BaseRequestSchemaAPI):
+    path = Resource_RequestPathSchema()
+
+
+class ResourceTypes_GET_ResponseBodySchema(BaseResponseBodySchema):
+    resource_name = colander.SchemaNode(
+        colander.String(),
+        description="Name of the specified resource."
+    )
+    resource_type = colander.SchemaNode(
+        colander.String(),
+        description="Type of the specified resource."
+    )
+    children_resource_types = ChildrenResourceTypesListSchema(
+        description="All resource types applicable as children under the specified resource. "
+    )
+    children_resource_allowed = colander.SchemaNode(
+        colander.Boolean(),
+        description="Indicates if the resource allows any children resources."
+    )
+    root_service_id = colander.SchemaNode(
+        colander.Integer(),
+        description="Resource tree root service identification number.",
+        default=colander.null,  # if no parent
+        missing=colander.drop  # if not returned (basic_info = True)
+    )
+    root_service_name = colander.SchemaNode(
+        colander.Integer(),
+        description="Resource tree root service name.",
+        default=colander.null,  # if no parent
+        missing=colander.drop  # if not returned (basic_info = True)
+    )
+    root_service_type = colander.SchemaNode(
+        colander.Integer(),
+        description="Resource tree root service type.",
+        default=colander.null,  # if no parent
+        missing=colander.drop  # if not returned (basic_info = True)
+    )
+
+
+class ResourceTypes_GET_OkResponseSchema(BaseResponseSchemaAPI):
+    description = "Get applicable children resource types under resource successful."
+    body = ResourceTypes_GET_ResponseBodySchema(code=HTTPOk.code, description=description)
 
 
 class ServiceResourcesBodySchema(ServiceBodySchema):
@@ -1602,6 +1649,22 @@ class ServiceTypeResources_GET_RequestSchema(BaseRequestSchemaAPI):
     path = ServiceType_RequestPathSchema()
 
 
+class ResourceTypesAllowed(colander.SequenceSchema):
+    description = "List of all allowed resource types at some level within the service hierarchy."
+    res_type_path = colander.SchemaNode(
+        colander.String(),
+        description="Allowed resource type under the service hierarchy."
+    )
+
+
+class ResourceStructuresAllowed(colander.SequenceSchema):
+    description = "List of allowed combinations of resource type hierarchical structures under the service."
+    res_type_path = colander.SchemaNode(
+        colander.String(),
+        description="Path-like resource type structure allowed under the service."
+    )
+
+
 class ServiceTypeResourceInfo(colander.MappingSchema):
     resource_type = colander.SchemaNode(
         colander.String(),
@@ -1609,8 +1672,10 @@ class ServiceTypeResourceInfo(colander.MappingSchema):
     )
     resource_child_allowed = colander.SchemaNode(
         colander.Boolean(),
-        description="Indicates if the resource type allows child resources."
+        description="Indicates if the service allows any child resources."
     )
+    resource_types_allowed = ResourceTypesAllowed()
+    resource_structure_allowed = ResourceStructuresAllowed()
     permission_names = PermissionNameListSchema(
         description="Permissions applicable to the specific resource type.",
         example=[Permission.READ.value, Permission.WRITE.value]
@@ -3274,6 +3339,15 @@ ResourcePermissions_GET_responses = {
     "403": Resource_MatchDictCheck_ForbiddenResponseSchema(),  # FIXME: https://github.com/Ouranosinc/Magpie/issues/359
     "404": Resource_MatchDictCheck_NotFoundResponseSchema(),
     "406": NotAcceptableResponseSchema(),
+    "422": UnprocessableEntityResponseSchema(),
+    "500": InternalServerErrorResponseSchema(),
+}
+ResourceTypes_GET_responses = {
+    "200": ResourceTypes_GET_OkResponseSchema(),
+    "400": Resource_MatchDictCheck_BadRequestResponseSchema(),
+    "401": UnauthorizedResponseSchema(),
+    "403": Resource_MatchDictCheck_ForbiddenResponseSchema(),
+    "404": Resource_MatchDictCheck_NotFoundResponseSchema(),
     "422": UnprocessableEntityResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }